requirements.txt
asgiref==3.3.1
backcall==0.2.0
decorator==4.4.2
Django==3.1.4
djangorestframework==3.12.2
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
Pygments==2.7.3
pytils==0.3
pytz==2020.4
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
